+++ CCApi.Extensions.DependencyInjection/AppExtensions.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using CCApi.Extensions.DependencyInjection;
 using CCApi.Extensions.DependencyInjection.Constants;
 using CCApi.Extensions.DependencyInjection.Database;
+using CCApi.Extensions.DependencyInjection.Utils;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
@@ -17,6 +18,9 @@ public static class AppExtensions
         Action<SwaggerUIOptions>? configureSwagger = null)
        where TDbContext : DappiDbContext
     {
+        var scope = app.Services.CreateScope();
+        var logger = scope.ServiceProvider.GetRequiredService<ILogger<TDbContext>>();
+
         await HandlePortCleanupAsync(app);
 
         if (app.Environment.IsDevelopment())
@@ -50,7 +54,6 @@ public static class AppExtensions
         app.MapControllers();
         app.UseStaticFiles();
 
-        var scope = app.Services.CreateScope();
         var db = scope.ServiceProvider.GetRequiredService<TDbContext>();
         try
         {
@@ -63,7 +66,6 @@ public static class AppExtensions
         // we can't migrate the schema changes, but we need to continue and just not publish un-published content-type changes.
         catch (InvalidOperationException e) when (e.Message.Contains("PendingModelChangesWarning"))
         {
-            var logger = scope.ServiceProvider.GetRequiredService<ILogger<TDbContext>>();
             logger.LogWarning(
                 "Unable to migrate schema changes due to pending model changes. Most probably you have models in draft state. {Message}",
                 e.Message);
@@ -75,9 +77,7 @@ public static class AppExtensions
     private static async Task HandlePortCleanupAsync(WebApplication app)
     {
         var args = Environment.GetCommandLineArgs();
-        var isRestart = args.Contains("--restart") ||
-                       Environment.GetEnvironmentVariable("DAPPI_RESTART") == "true" ||
-                       Environment.GetEnvironmentVariable("DAPPI_MIGRATION_RESTART") == "true";
+        var isRestart = args.Contains("--restart");
 
         if (isRestart)
         {
@@ -89,125 +89,11 @@ public static class AppExtensions
 
         if (!string.IsNullOrEmpty(urls))
         {
-            await KillProcessesUsingUrls(urls);
+            await ProcessUtils.KillProcessesUsingUrls(urls);
         }
     }
 
-    private static async Task KillProcessesUsingUrls(string urls)
-    {
-        var ports = new HashSet<int>();
-        var urlList = urls.Split(';');
-
-        foreach (var url in urlList)
-        {
-            if (Uri.TryCreate(url.Trim(), UriKind.Absolute, out var uri))
-            {
-                ports.Add(uri.Port);
-            }
-        }
-
-        foreach (var port in ports)
-        {
-            await KillProcessUsingPort(port);
-        }
-    }
-
-    private static async Task KillProcessUsingPort(int port)
-    {
-        try
-        {
-            Console.WriteLine($"Checking for processes using port {port}...");
-
-            var pids = await GetProcessIdsUsingPort(port);
-
-            if (pids.Any())
-            {
-                foreach (var pid in pids)
-                {
-                    if (pid != Environment.ProcessId)
-                    {
-                        await KillProcessById(pid, port);
-                    }
-                    else
-                    {
-                        Console.WriteLine($"Skipping current process {pid} on port {port}");
-                    }
-                }
-            }
-            else
-            {
-                Console.WriteLine($"No processes found using port {port}");
-            }
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error checking port {port}: {ex.Message}");
-        }
-    }
-
-    private static async Task<List<int>> GetProcessIdsUsingPort(int port)
-    {
-        var pids = new List<int>();
-
-        var processInfo = new ProcessStartInfo
-        {
-            FileName = OperatingSystem.IsWindows() ? "cmd.exe" : "bash",
-            Arguments = OperatingSystem.IsWindows()
-                ? $"/c netstat -ano | findstr :{port}"
-                : $"-c \"lsof -ti:{port}\"",
-            UseShellExecute = false,
-            RedirectStandardOutput = true,
-            CreateNoWindow = true
-        };
-
-        using var process = Process.Start(processInfo);
-        var output = await process.StandardOutput.ReadToEndAsync();
-        await process.WaitForExitAsync();
-
-        if (string.IsNullOrEmpty(output)) return pids;
-
-        if (OperatingSystem.IsWindows())
-        {
-            var lines = output.Split('\n', StringSplitOptions.RemoveEmptyEntries);
-            foreach (var line in lines)
-            {
-                var parts = line.Trim().Split(' ', StringSplitOptions.RemoveEmptyEntries);
-                if (parts.Length > 0 && int.TryParse(parts[^1], out int pid))
-                {
-                    pids.Add(pid);
-                }
-            }
-        }
-        else
-        {
-            var pidStrings = output.Split('\n', StringSplitOptions.RemoveEmptyEntries);
-            foreach (var pidStr in pidStrings)
-            {
-                if (int.TryParse(pidStr.Trim(), out int pid))
-                {
-                    pids.Add(pid);
-                }
-            }
-        }
-
-        return pids.Distinct().ToList();
-    }
-
-    private static async Task KillProcessById(int pid, int port)
-    {
-        try
-        {
-            var processToKill = Process.GetProcessById(pid);
-            processToKill.Kill();
-            await processToKill.WaitForExitAsync();
-            Console.WriteLine($"✓ Killed process {pid} using port {port}");
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Failed to kill process {pid}: {ex.Message}");
-        }
-    }
-
+ 
     public static async Task SeedRolesAndUsersAsync<TUser, TRole>(this IServiceProvider serviceProvider)
         where TUser : IdentityUser, new()
         where TRole : IdentityRole, new()
